Split widget into separated files.
