+++ lib/detailpage.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+
+class ItemDetailPage extends StatefulWidget {
+  const ItemDetailPage({super.key, required this.item});
+
+  final String item;
+
+  @override
+  State<ItemDetailPage> createState() => _ItemDetailPageState();
+}
+
+class _ItemDetailPageState extends State<ItemDetailPage> {
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      children: [SafeArea(child: Text(widget.item))],
+    );
+  }
+}
